titiler_service/requirements.txt
titiler.application==2.0.0
uvicorn